.setup/requirements.txt
alembic==1.4.1
aplus==0.11.0
arrow==0.15.5
astropy==4.0.1.post1
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
binaryornot==0.4.4
bleach==3.1.3
blinker==1.4
bokeh==1.4.0
boto3==1.12.36
botocore==1.15.36
bqplot==0.12.6
branca==0.3.1
cachetools==4.0.0
certifi==2019.11.28
certipy==0.1.3
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
conda==4.8.2
conda-package-handling==1.6.0
cookiecutter==1.7.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distributed==2.12.0
docutils==0.15.2
entrypoints==0.3
fsspec==0.6.2
future==0.18.2
h5py==2.10.0
HeapDict==1.0.1
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
ipydatawidgets==4.0.1
ipykernel==5.1.4
ipyleaflet==0.12.3
ipympl==0.5.3
ipython==7.13.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.3.0
ipyvuetify==1.2.2
ipywebrtc==0.5.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jinja2-time==0.2.0
jmespath==0.9.5
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyter-telemetry==0.0.5
jupyterhub==1.1.0
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.1.0
llvmlite==0.31.0
locket==0.2.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mpmath==1.1.0
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nodejs==0.1.1
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
oauthlib==3.0.1
olefile==0.46
optional-django==0.1.0
packaging==20.1
pamela==1.0.0
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
poyo==0.5.0
progressbar2==3.50.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
pyarrow==0.16.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.5
Pygments==2.6.1
PyJWT==1.7.1
PyOpenGL==3.1.5
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==0.1.11
python-utils==2.4.0
pythreejs==2.2.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
ruamel-yaml==0.15.80
ruamel.yaml.clib==0.2.0
s3fs==0.2.2
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.4
SQLAlchemy==1.3.15
statsmodels==0.11.1
sympy==1.5.1
tabulate==0.8.7
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
traittypes==0.2.1
urllib3==1.25.7
vaex-arrow==0.4.2
vaex-astro==0.6.1
vaex-core==1.5.0
vaex-hdf5==0.5.6
vaex-jupyter==0.4.1
vaex-server==0.2.1
vaex-ui==0.3.0
vaex-viz==0.3.8
vincent==0.4.4
wcwidth==0.1.8
webencodings==0.5.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
xarray==0.15.1
xlrd==1.2.0
zict==2.0.0
zipp==3.1.0